Ending Beta
Test with a larger playerbase went well. Fixed some bugs and glitches. Enhanced and clarified parts of the documentation. Added a command for questwrite to check for loot-messages from the chat log instead of spoken keywords for individual sections. Ending Beta. Switching to version numbering that marks year and month, dropping the day.