Many small changes and new functions in QuestWrite
Added a simple fishing contest, documentation within the .qwt. Switched timestamps in QW to a format native to AHK as there are no math operations for other formats. Added a history for players sections and times that is kept in memory and is exported with CSV export. Added qwt_timegap command to .qwt files to tarpit clowns who try to spam triggers. Added qwt_invite command to .qwt files to automatically invite participants into group/fellow/raid for having loot messages in the chatlog and being able to check if items are looted for a quest,
Updated the userdoc accordingly.